requirements.txt
boto3==1.24.42
Flask==2.1.3
redis==4.3.4
requests==2.28.1
Unidecode==1.3.4
python-dotenv==0.20.0
